fix: round stock qty before post to wc
